Extract engine contract types
