+++ packages/column-live-view-engine/src/engine-errors.ts
@@ -1,0 +1,37 @@
+import { Schema } from "effect";
+import type { InvalidQueryError } from "./raw-query-compiler.js";
+
+export class InvalidTopicError extends Schema.TaggedErrorClass<InvalidTopicError>()(
+  "InvalidTopicError",
+  {
+    topic: Schema.String,
+    message: Schema.String,
+  },
+) {}
+
+export class InvalidRowError extends Schema.TaggedErrorClass<InvalidRowError>()("InvalidRowError", {
+  topic: Schema.String,
+  message: Schema.String,
+}) {}
+
+export class UnsupportedQueryError extends Schema.TaggedErrorClass<UnsupportedQueryError>()(
+  "UnsupportedQueryError",
+  {
+    topic: Schema.String,
+    message: Schema.String,
+  },
+) {}
+
+export class EngineClosedError extends Schema.TaggedErrorClass<EngineClosedError>()(
+  "EngineClosedError",
+  {
+    message: Schema.String,
+  },
+) {}
+
+export type ColumnLiveViewEngineError =
+  | InvalidTopicError
+  | InvalidRowError
+  | UnsupportedQueryError
+  | InvalidQueryError
+  | EngineClosedError;
